refactor(sessions): extract ctrl to make it testable
